+++ src/utils/nanvixd/src/sandbox/multi_process/mod.rs
@@ -6,4 +6,4 @@
 //==================================================================================================
 
 pub mod linuxd;
-pub mod microvm;
+pub mod uservm;
